Implemented live updates on crud operations using socket

+++ server/controllers/taskController.js
@@ -1,11 +1,15 @@
 import {
+    emitTasks
+} from "../index.js";
+import {
     taskModel
 } from "../models/taskModel.js"
 
 const getTaskController = async (req, res, next) => {
-    console.log("Get taks...")
+    console.log("Get tasks...")
     const userId = req.userId ? req.userId : null;
     try {
+
         if (!userId) {
             return res.status(400).json({
                 message: "User ID is required."
@@ -24,19 +28,29 @@ const getTaskController = async (req, res, next) => {
 
 const addNewTaskController = async (req, res, next) => {
     const {
-        taskDatas
+        taskData
     } = req.body;
     try {
-        const userId = req.user ? req.user.id : null;
+        const {
+            title,
+            dueDate
+        } = taskData
+        const userId = req.userId ? req.userId : null;
         if (!userId) {
             return res.status(400).json({
-                message: "User ID is required."
+                message: "User is required."
             })
         };
         const newTask = await taskModel.create({
             userId,
-            taskDatas
+            title,
+            dueDate
         });
+        const stats = await statusCalculator(userId);
+        emitTasks("taskAdded", {
+            newTask,
+            stats
+        })
         return res.status(201).json({
             message: "Task created successfully",
             task: newTask
@@ -49,16 +63,22 @@ const editTaskController = async (req, res, next) => {
     const {
         taskDatas
     } = req.body;
+    const userId = req.userId;
     try {
+        const {
+            title,
+            dueDate
+        } = taskDatas;
         const taskId = req.params.id;
-        if (!taskId) {
+        if (!taskId || !userId) {
             return res.status(400).json({
-                message: "Task ID is required."
+                message: "ID is required."
             })
         };
         const updatedTask = await taskModel.findByIdAndUpdate(
             taskId, {
-                taskDatas
+                title,
+                dueDate
             }, {
                 new: true
             });
@@ -67,6 +87,11 @@ const editTaskController = async (req, res, next) => {
                 message: "Task not found."
             });
         }
+        const stats = await statusCalculator(userId);
+        emitTasks("taskUpdated", {
+            updatedTask,
+            stats
+        });
         return res.status(200).json({
             message: "Task udated successfully",
             task: updatedTask
@@ -76,8 +101,15 @@ const editTaskController = async (req, res, next) => {
     }
 }
 const deleteTaskController = async (req, res, next) => {
+    console.log("Delete task....")
     const taskId = req.params.id;
+    const userId = req.userId;
     try {
+        if (!userId) {
+            return res.status(401).json({
+                message: "Unauthorized."
+            })
+        }
         const task = await taskModel.findByIdAndDelete(
             taskId
         );
@@ -86,6 +118,11 @@ const deleteTaskController = async (req, res, next) => {
                 message: "Task not found."
             });
         }
+        const stats = await statusCalculator(userId);
+        emitTasks("taskDeleted", {
+            task,
+            stats
+        });
         return res.status(200).json({
             message: "Task deleted successfully",
             task
@@ -94,11 +131,87 @@ const deleteTaskController = async (req, res, next) => {
         next(error)
     }
 }
+const updateTaskStatusController = async (req, res, next) => {
+    const taskId = req.params.id;
+    const userId = req.userId;
+    try {
+        if (!taskId || !userId) {
+            return res.status(400).json({
+                message: "Id is required."
+            })
+        }
+        const task = await taskModel.findById(taskId);
+        if (!task) {
+            return res.status(404).json({
+                message: "Task not found."
+            });
+        }
+        task.completed = !task.completed
+        const updatedTask = await task.save();
+        const stats = await statusCalculator(userId);
+        emitTasks("taskUpdated", {
+            updatedTask,
+            stats
+        });
+        return res.status(200).json({
+            message: "Task updated..",
+            task
+        });
+    } catch (error) {
+        next(error)
+    }
+}
+const getTaskStatsController = async (req, res, next) => {
+    const userId = req.userId;
+    try {
+        if (!userId) {
+            return res.status(400).json({
+                message: "Task id is required."
+            })
+        }
+        const {
+            overdueTasks,
+            completedTasks,
+            totalTasks
+        } = await statusCalculator(userId);
+        return res.status(200).json({
+            message: "Task updated..",
+            overdueTasks,
+            completedTasks,
+            totalTasks
+        });
+    } catch (error) {
+        next(error)
+    }
+}
+
+const statusCalculator = async (userId) => {
+    try {
+        const tasks = await taskModel.find({
+            userId
+        });
+        const completedTasks = tasks.filter(task => task.completed);
+        const totalTasks = tasks.length;
+
+        const currentDate = new Date();
+        const overdueTasks = tasks.filter(task => task.dueDate && new Date(task.dueDate) < currentDate);
+
+        return {
+            totalTasks,
+            completedTasks: completedTasks.length,
+            overdueTasks: overdueTasks.length
+        };
+    } catch (error) {
+        throw error;
+    }
+};
 
 
 export {
     addNewTaskController,
     getTaskController,
     editTaskController,
-    deleteTaskController
+    deleteTaskController,
+    updateTaskStatusController,
+    getTaskStatsController
 }
